Add restructure to project [Delivers-#163286622]
