+++ api/Helpers.py
@@ -1,4 +1,5 @@
 from api.models.user import users
+from flask import jsonify
 from api.models.incident import incidents
 from werkzeug.security import check_password_hash
 
@@ -39,3 +40,5 @@ def check_status():
             if redflag['status'] != 'draft':
                 return True
         return False
+
+
